requirements.txt
Django==1.11
Pillow==4.1.0
django-ckeditor==5.2.2
